Revert "Feature/9 (#28)" (#32)
This reverts commit 18c9579ecf0baf30f5f5d7857912467bb77f131e.
