+++ backend/src/main/java/com/java/NBE4_5_1_8/domain/orderinfo/repository/OrderInfoRepository.java
@@ -6,5 +6,4 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface OrderInfoRepository extends JpaRepository<OrderInfo, Long> {
-    OrderInfo findByMemberEmail(String memberEmail);
 }
