+++ backend/src/main/java/com/java/NBE4_5_1_8/domain/orderitem/OrderItem.java
@@ -1,0 +1,30 @@
+package com.java.NBE4_5_1_8.domain.orderitem;
+
+import com.java.NBE4_5_1_8.domain.item.entity.Item;
+import com.java.NBE4_5_1_8.domain.orderinfo.entity.OrderInfo;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+public class OrderItem {
+
+    @Id
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name = "item_id")
+    private Item item;
+
+    @ManyToOne
+    @JoinColumn(name = "order_id")
+    private OrderInfo orderInfo;
+
+    private int orderPrice;
+    private int quantity;
+}
